# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CurrentWorkloadMetrics:
    properties:
      cpu_limit_cores:
        type: number
      cpu_projected_monthly_waste_usd:
        type: number
      cpu_requested_cores:
        type: number
      cpu_used_cores:
        type: number
      cpu_waste_ratio:
        type: number
      mem_waste_ratio:
        type: number
      memory_limit_bytes:
        type: number
      memory_projected_monthly_waste_usd:
        type: number
      memory_requested_bytes:
        type: number
      memory_used_bytes:
        type: number
      projected_monthly_waste_usd:
        type: number
      timestamp:
        type: string
    type: object
  handlers.HistoricalDetailPoint:
    properties:
      consolidation_id:
        type: integer
      cpu_limit_cores:
        type: number
      cpu_projected_monthly_waste_usd:
        type: number
      cpu_requested_cores:
        type: number
      cpu_used_cores:
        type: number
      cpu_waste_ratio:
        type: number
      mem_waste_ratio:
        type: number
      memory_limit_bytes:
        type: number
      memory_projected_monthly_waste_usd:
        type: number
      memory_requested_bytes:
        type: number
      memory_used_bytes:
        type: number
      projected_monthly_waste_usd:
        type: number
      timestamp:
        type: string
    type: object
  handlers.HistoricalWorkloadSummary:
    properties:
      avg_cpu_projected_monthly_waste_usd:
        type: number
      avg_cpu_requested_cores:
        type: number
      avg_cpu_used_cores:
        type: number
      avg_cpu_waste_ratio:
        type: number
      avg_mem_waste_ratio:
        type: number
      avg_memory_projected_monthly_waste_usd:
        type: number
      avg_memory_requested_bytes:
        type: number
      avg_memory_used_bytes:
        type: number
      avg_projected_monthly_waste_usd:
        type: number
      last_10_consolidations_count:
        type: integer
      points:
        items:
          $ref: '#/definitions/handlers.HistoricalDetailPoint'
        type: array
    type: object
...
  handlers.WorkloadDetailsResponse:
    properties:
      container:
        type: string
      current:
        $ref: '#/definitions/handlers.CurrentWorkloadMetrics'
      historical:
        $ref: '#/definitions/handlers.HistoricalWorkloadSummary'
      namespace:
        type: string
      pod:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Suporte Go-Analyze
  description: Esta é a API do analisador de recursos e custos de workloads Kubernetes
    (go-analyze).
  title: Go Analyze API
  version: "1.0"
paths:
...
  /api/v1/experiment/reset:
    post:
      consumes:
      - application/json
      description: Limpa as tabelas de histórico do banco de dados e zera o estado
        interno dos trackers de notificação para uma nova rodada de teste.
      parameters:
      - description: Bearer token experimental
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Resetar dados do experimento
      tags:
      - experiment
...
  /api/v1/workloads/details:
    get:
      consumes:
      - application/json
      description: Retorna as métricas atuais de execução em tempo real e um histórico
        das consolidações registradas para o workload. O parâmetro `pod` é tratado
        como prefixo (LIKE) quando não contém o caractere %.
      parameters:
      - description: Namespace do Kubernetes
        in: query
        name: namespace
        required: true
        type: string
      - description: Nome do Container no Kubernetes
        in: query
        name: container
        required: true
        type: string
      - description: Prefixo do Pod no Kubernetes (LIKE)
        in: query
        name: pod
        type: string
      - description: Janela de tempo para o cálculo atual
        in: query
        name: window
        type: string
      - description: 'Período rápido para busca histórica (ex: 1hr, 5d, 10d)'
        in: query
        name: period
        type: string
      - description: 'Data inicial para histórico (ex: RFC3339)'
        in: query
        name: start_at
        type: string
      - description: Data final para histórico
        in: query
        name: end_at
        type: string
      - description: 'Número máximo de pontos históricos a retornar (padrão: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.WorkloadDetailsResponse'
        "400":
          description: Namespace e container são parâmetros obrigatórios ou datas
            inválidas
          schema:
            additionalProperties: true
            type: object
        "405":
          description: Método HTTP não permitido
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro ao carregar os dados históricos
          schema:
            additionalProperties: true
            type: object
      summary: Obter detalhes de um workload específico
      tags:
      - workloads
swagger: "2.0"